pyproject.toml
[project]
name = "business-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "dotenv>=0.9.9",
    "google-adk>=0.3.0",
    "mcp>=1.6.0",
    "youtube-transcript-api>=1.0.3",
]
